poc/requirements.txt
openai>=1.12.0
weaviate-client>=4.5.0
python-dotenv>=1.0.0
